task_app/pyproject.toml
[tool.poetry]
name = "task-app"
version = "0.1.0"
description = ""
authors = ["mukhtar-ahmed <kh.mukhtar60@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
sqlalchemy = "^2.0.41"
pydantic = "^2.11.4"
passlib = "^1.7.4"
bcrypt = "4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
